\name{flexCWM-package}
\alias{flexCWM-package}
\alias{flexCWM}
\docType{package}
\title{flexCWM - Flexible Cluster Weighted Modeling}
\description{
This package allows for Flexible Cluster Weighted Modeling.
}
\details{
\tabular{ll}{
Package: \tab flexCWM\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-07-06\cr
License: \tab GNU-2\cr
}
}
\references{
Ingrassia, S., Minotti, S. C., and Vittadini, G. (2012). Local statistical modeling via the cluster-weighted approach with elliptical distributions. \emph{Journal of Classification}, \bold{29}(3), 363-401.

Ingrassia, S., Minotti, S. C., Punzo, A., and Vittadini, G. (2012). Generalized linear Gaussian cluster-weighted
modeling. \emph{arXiv.org} e-print \bold{1211.1171}, available at: http://arxiv.org/abs/1211.1171.

}
\author{
Mazza A., Punzo A., Ingrassia S.

Maintainer: Mazza Angelo <a.mazza@unict.it>
}
%%\note{
%%  ~~further notes~~
%%}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{glmcwm}}, \code{\link{tourism}}
}