\name{tourism}
\docType{data}
\alias{tourism}
\title{tourism}
\description{
The \code{tourism} dataset consists of 180 monthly data concerning tourist overnights (variable \code{overnights}) and attendance at museums and
monuments in Italy (variable \code{attendance}) over the 15-year period spanning from January 1996 to December 2010.
}
\usage{data(tourism)}

\format{A dataset}
\source{\url{http://www.robertocellini.it/doc/master_specializzazione/Cellini-Cuccia_ApEc2013_data1996-2010.pdf}}
\references{
Cellini, R., & Cuccia, T. (2013). Museum and monument attendance and tourism flow: A time series analysis approach. \emph{Applied Economics}, \bold{45}, 3473-3482.

Ingrassia, S., Minotti, S. C., and Punzo, A. (2013). Model-based clustering via linear cluster-weighted models.
\emph{Computational Statistics and Data Analysis}, DOI: 10.1016/j.csda.2013.02.012.

}
\seealso{
\code{\link{flexCWM-package}}, \code{\link{glmcwm}}
}
\keyword{datasets}