\name{flexCWM-package}
\alias{flexCWM-package}
\docType{package}
\title{flexCWM - Flexible Cluster Weighted Modeling}
\description{
Allows for maximum likelihood fitting of cluster-weighted models, a class of mixtures of regression models with random covariates. 
}
\details{
\tabular{ll}{
Package: \tab CWM\cr
Type: \tab Package\cr
Version: \tab 1.7\cr
Date: \tab 2017-02-14\cr
License: \tab GNU-2\cr
}
}
\references{
Mazza, A., Ingrassia, S., and Punzo, A. (2018). {flexCWM}: A Flexible Framework for Cluster-Weighted Models.  \emph{Journal of Statistical Software}, \bold{86}(2), 1-30.

Ingrassia, S., Minotti, S. C., and Vittadini, G. (2012). Local Statistical Modeling via the Cluster-Weighted Approach with Elliptical Distributions. \emph{Journal of Classification}, \bold{29}(3), 363-401.

Ingrassia, S., Minotti, S. C., and Punzo, A. (2014). Model-based clustering via linear cluster-weighted models. \emph{Computational Statistics and Data Analysis}, \bold{71}, 159-182.

Ingrassia, S., Punzo, A., and Vittadini, G. (2015). The Generalized Linear Mixed Cluster-Weighted Model. \emph{Journal of Classification}, \bold{32}(forthcoming)

Punzo, A. (2014). Flexible Mixture Modeling with the Polynomial Gaussian Cluster-Weighted Model. \emph{Statistical Modelling}, \bold{14}(3), 257-291.

%Punzo, A. and Ingrassia, S. (2015a). Clustering Bivariate Mixed-Type Data via the Cluster-Weighted Model. Computational Statistics, ..(..), ..-..

%Punzo, A. and Ingrassia, S. (2015b). Parsimonious Generalized Linear Gaussian Cluster-Weighted Models. In T. Minerva, I. Morlini, and F. Palumbo (eds.), \emph{Statistical Models for Data Analysis}, Studies in Classification, Data Analysis and Knowledge Organization, pp. ..-.. Springer International Publishing, Switzerland.
}
\author{
Mazza A., Punzo A., Ingrassia S.

Maintainer: Mazza Angelo <a.mazza@unict.it>
}
%%\note{
%%  ~~further notes~~
%%}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{cwm}}
}